requirements.txt
streamlit
transformers
torch
pillow
openai